refactored dashboard to favorites
