+++ app/src/main/java/com/example/mom_pops/ui/favorites/FavoritesViewModel.java
@@ -1,16 +1,16 @@
-package com.example.mom_pops.ui.dashboard;
+package com.example.mom_pops.ui.favorites;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class FavoritesViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public FavoritesViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("Favorites Page");
     }
 
     public LiveData<String> getText() {
